# Refactor strmatcher.MphMatcherGroup