As per description in https://github.com/v2fly/v2ray-core/discussions/1275#discussioncomment-1559883:

> `MphMatcherGroup` uses an algorithm [Hash, displace, and compress](http://cmph.sourceforge.net/papers/esa09.pdf) to construct a minimal perfect hash table. It involves `1. a normal hash function`, and `2. a series of hash functions independent to each others`.
> 
> This is how this `MphMatcherGroup` works:
> 
> #### First level hash: RollingHash
> Rolling hash is chosen here because we can easily calculate the hash of a full domain and all its subdomains.
> 
> We obtain a seed from the result of first level hash.
> 
> #### Second level hash: MemHash
> One of second level functions is selected according to the seed obtained from first level hash.
> 
> This is done by choosing golang runtime's `memhash`, because `memhash` requires a seed as argument, and with different seeds, `memhash` behaves exactly like distinct hash functions independent to each other.
> 
> To use `memhash`, current implementation of `MphMatcherGroup` copies source code (unoptimized version) from golang runtime. However, we could directly reference the runtime 'simplementation using compiler directive `go:linkname`:
> 
> ```
> //go:noescape
> //go:linkname memhash runtime.memhash
> func memhash(p unsafe.Pointer, h, s uintptr) uintptr
> ```
> 
> While avoiding duplicate code, the runtime's memhash implementation could also benefit from hardware optimizations like AES if supported. Current `MphMatcherGroup` unconditionally uses the unoptimized fallback version.