@Vigilans 2021-10-31
### Benchmark

#### Domain Matcher

* Original `MatchAny`: 34.95 ns/op
* New `MatchAny`: 32.49 ns/op
* New `Match`: 40.25 ns/op

#### Full Matcher

* Original `MatchAny`: 34.67 ns/op
* New `MatchAny`: 26.86 ns/op
* New `Match`: 40.43 ns/op

Because only Domain matcher needs to lookup the mph table every time a dot '.' is met, so `ACAutomationMatcherGroup` could actually be further optimized according to what matchers are stored in it and what type the upstream wants to match. This will be implemented in `EtldIndexMatcher`'s PR.